requirements.txt
Flask
stravalib
python-dotenv 